Moved Throttler to ihmc-commons
