+++ ihmc-communication/src/main/java/us/ihmc/communication/property/ROS2StoredPropertySet.java
@@ -2,7 +2,7 @@ package us.ihmc.communication.property;
 
 import us.ihmc.communication.ros2.ROS2PublishSubscribeAPI;
 import us.ihmc.tools.property.StoredPropertySetBasics;
-import us.ihmc.tools.thread.Throttler;
+import us.ihmc.commons.thread.Throttler;
 
 /**
  * ROS 2 enabled, synced, interprocess stored property set. It allows external process
