+++ ihmc-communication/src/main/java/us/ihmc/communication/ros2/ROS2Heartbeat.java
@@ -1,14 +1,14 @@
 package us.ihmc.communication.ros2;
 
 import std_msgs.msg.dds.Empty;
+import us.ihmc.commons.UnitConversions;
 import us.ihmc.commons.exception.DefaultExceptionHandler;
 import us.ihmc.commons.exception.ExceptionTools;
 import us.ihmc.commons.thread.ThreadTools;
-import us.ihmc.ros2.ROS2PublisherBasics;
+import us.ihmc.commons.thread.Throttler;
 import us.ihmc.ros2.ROS2NodeInterface;
+import us.ihmc.ros2.ROS2PublisherBasics;
 import us.ihmc.ros2.ROS2Topic;
-import us.ihmc.commons.UnitConversions;
-import us.ihmc.tools.thread.Throttler;
 
 /**
  * Use this class to indicate to other things on the network that something is active.
