+++ ihmc-communication/src/main/java/us/ihmc/communication/ros2/ROS2HeartbeatMonitor.java
@@ -8,7 +8,7 @@ import us.ihmc.commons.thread.TypedNotification;
 import us.ihmc.ros2.ROS2Input;
 import us.ihmc.ros2.ROS2Topic;
 import us.ihmc.tools.Timer;
-import us.ihmc.tools.thread.Throttler;
+import us.ihmc.commons.thread.Throttler;
 
 import java.util.function.Consumer;
