+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/logging/FFMPEGVideoPlaybackManager.java
@@ -8,7 +8,7 @@ import us.ihmc.log.LogTools;
 import us.ihmc.perception.BytedecoImage;
 import us.ihmc.perception.ffmpeg.FFmpegTools;
 import us.ihmc.perception.logging.PerceptionLoggerConstants;
-import us.ihmc.tools.thread.Throttler;
+import us.ihmc.commons.thread.Throttler;
 
 public class FFMPEGVideoPlaybackManager
 {
