+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/perception/RDXBlackflyCalibrationSuite.java
@@ -37,7 +37,7 @@ import us.ihmc.rdx.ui.interactable.RDXInteractableBlackflyFujinon;
 import us.ihmc.tools.thread.MissingThreadTools;
 import us.ihmc.tools.thread.ResettableExceptionHandlingExecutorService;
 import us.ihmc.tools.thread.SwapReference;
-import us.ihmc.tools.thread.Throttler;
+import us.ihmc.commons.thread.Throttler;
 
 import java.util.ArrayList;
