+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/ui/footstepPlanner/RDXFootstepPlanning.java
@@ -32,7 +32,7 @@ import us.ihmc.robotics.robotSide.RobotSide;
 import us.ihmc.sensorProcessing.heightMap.HeightMapData;
 import us.ihmc.tools.thread.MissingThreadTools;
 import us.ihmc.tools.thread.ResettableExceptionHandlingExecutorService;
-import us.ihmc.tools.thread.Throttler;
+import us.ihmc.commons.thread.Throttler;
 
 import java.util.ArrayList;
