+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/ui/graphics/RDXMessageSizeReadout.java
@@ -2,7 +2,7 @@ package us.ihmc.rdx.ui.graphics;
 
 import imgui.internal.ImGui;
 import us.ihmc.commons.FormattingTools;
-import us.ihmc.tools.thread.Throttler;
+import us.ihmc.commons.thread.Throttler;
 
 /**
  * Keep track of the size of the incoming messages while slowing it down
