+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/ui/interactable/RDXNeckPitchSlider.java
@@ -4,10 +4,10 @@ import controller_msgs.msg.dds.RobotConfigurationData;
 import imgui.internal.ImGui;
 import us.ihmc.avatar.drcRobot.ROS2SyncedRobotModel;
 import us.ihmc.avatar.ros2.ROS2ControllerHelper;
+import us.ihmc.commons.UnitConversions;
+import us.ihmc.commons.thread.Throttler;
 import us.ihmc.rdx.imgui.ImGuiUniqueLabelMap;
 import us.ihmc.rdx.ui.teleoperation.RDXTeleoperationParameters;
-import us.ihmc.commons.UnitConversions;
-import us.ihmc.tools.thread.Throttler;
 
 public class RDXNeckPitchSlider
 {
