+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/ui/interactable/RDXPelvisHeightSlider.java
@@ -16,7 +16,7 @@ import us.ihmc.humanoidRobotics.frames.HumanoidReferenceFrames;
 import us.ihmc.log.LogTools;
 import us.ihmc.commons.UnitConversions;
 import us.ihmc.tools.string.StringTools;
-import us.ihmc.tools.thread.Throttler;
+import us.ihmc.commons.thread.Throttler;
 
 public class RDXPelvisHeightSlider
 {
