+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/ui/interactable/RDXSakeHandWidgets.java
@@ -8,7 +8,9 @@ import us.ihmc.avatar.sakeGripper.ROS2SakeHandStatus;
 import us.ihmc.avatar.sakeGripper.SakeHandParameters;
 import us.ihmc.avatar.sakeGripper.SakeHandPreset;
 import us.ihmc.behaviors.tools.CommunicationHelper;
+import us.ihmc.commons.UnitConversions;
 import us.ihmc.commons.thread.Notification;
+import us.ihmc.commons.thread.Throttler;
 import us.ihmc.communication.SakeHandAPI;
 import us.ihmc.log.LogTools;
 import us.ihmc.rdx.imgui.ImGuiFlashingText;
@@ -19,8 +21,6 @@ import us.ihmc.rdx.imgui.ImGuiUniqueLabelMap;
 import us.ihmc.robotics.EuclidCoreMissingTools;
 import us.ihmc.robotics.robotSide.RobotSide;
 import us.ihmc.tools.Timer;
-import us.ihmc.commons.UnitConversions;
-import us.ihmc.tools.thread.Throttler;
 
 public class RDXSakeHandWidgets
 {
