+++ ihmc-java-toolkit/src/main/java/us/ihmc/tools/thread/MissingThreadTools.java
@@ -2,6 +2,7 @@ package us.ihmc.tools.thread;
 
 import us.ihmc.commons.Conversions;
 import us.ihmc.commons.thread.ThreadTools;
+import us.ihmc.commons.thread.Throttler;
 
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
