+++ ihmc-perception/src/main/java/us/ihmc/sensors/RealsenseColorDepthLogger.java
@@ -4,9 +4,11 @@ import org.bytedeco.javacpp.BytePointer;
 import org.bytedeco.opencv.global.opencv_core;
 import org.bytedeco.opencv.opencv_core.Mat;
 import perception_msgs.msg.dds.ImageMessage;
+import us.ihmc.commons.UnitConversions;
 import us.ihmc.commons.exception.DefaultExceptionHandler;
 import us.ihmc.commons.nio.FileTools;
 import us.ihmc.commons.thread.ThreadTools;
+import us.ihmc.commons.thread.Throttler;
 import us.ihmc.euclid.referenceFrame.FramePose3D;
 import us.ihmc.euclid.referenceFrame.ReferenceFrame;
 import us.ihmc.perception.MutableBytePointer;
@@ -19,8 +21,6 @@ import us.ihmc.perception.realsense.RealsenseDeviceManager;
 import us.ihmc.perception.tools.PerceptionDebugTools;
 import us.ihmc.perception.tools.PerceptionMessageTools;
 import us.ihmc.tools.IHMCCommonPaths;
-import us.ihmc.commons.UnitConversions;
-import us.ihmc.tools.thread.Throttler;
 
 import java.text.SimpleDateFormat;
 import java.time.Instant;
