+++ ihmc-perception/src/test/java/us/ihmc/perception/streaming/ROS2StreamStatusMonitorTest.java
@@ -12,7 +12,7 @@ import us.ihmc.communication.ros2.ROS2Helper;
 import us.ihmc.pubsub.DomainFactory.PubSubImplementation;
 import us.ihmc.ros2.ROS2Node;
 import us.ihmc.ros2.ROS2Topic;
-import us.ihmc.tools.thread.Throttler;
+import us.ihmc.commons.thread.Throttler;
 
 import java.net.InetSocketAddress;
