+++ ihmc-perception/src/test/java/us/ihmc/perception/streaming/SRTStreamerSubscriberTest.java
@@ -24,7 +24,7 @@ import us.ihmc.perception.opencv.OpenCVTools;
 import us.ihmc.pubsub.DomainFactory.PubSubImplementation;
 import us.ihmc.ros2.ROS2Node;
 import us.ihmc.ros2.ROS2Topic;
-import us.ihmc.tools.thread.Throttler;
+import us.ihmc.commons.thread.Throttler;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
